requirements.txt
google-genai
pandas
pytest
streamlit
